# src/internal/api/specdata/openapi.yaml
openapi: 3.0.3
info:
  title: NeuronAgent API
  description: |
    REST API for NeuronAgent - AI agent runtime system providing agent state management,
    session handling, message processing, tool execution, and streaming capabilities.
  version: 2.0.0
  contact:
    name: NeuronDB Support
    email: support@neurondb.ai
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.neurondb.ai/v1
    description: Production server
...
paths:
...
  /agents:
    get:
      tags:
        - Agents
      summary: List agents
      description: List all agents, optionally filtered by search query
      operationId: listAgents
      parameters:
        - name: search
          in: query
          description: Search query to filter agents
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Agents
      summary: Create agent
      description: Create a new agent
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /agents/{id}:
    get:
      tags:
        - Agents
      summary: Get agent
      description: Get agent by ID
      operationId: getAgent
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Agents
      summary: Update agent
      description: Update an existing agent
      operationId: updateAgent
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: Agent updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Agents
      summary: Delete agent
      description: Delete an agent by ID
      operationId: deleteAgent
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '204':
          description: Agent deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sessions:
    post:
      tags:
        - Sessions
      summary: Create session
      description: Create a new session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sessions/{id}:
    get:
      tags:
        - Sessions
      summary: Get session
      description: Get session by ID
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Sessions
      summary: Update session
      description: Update an existing session
      operationId: updateSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSessionRequest'
      responses:
        '200':
          description: Session updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Sessions
      summary: Delete session
      description: Delete a session by ID
      operationId: deleteSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '204':
          description: Session deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sessions/{session_id}/messages:
    get:
      tags:
        - Messages
      summary: Get messages
      description: Get messages for a session
      operationId: getMessages
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - name: limit
          in: query
          description: Maximum number of messages to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          description: Number of messages to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Messages
      summary: Send message
      description: Send a message to a session
      operationId: sendMessage
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: Message processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tools:
    get:
      tags:
        - Tools
      summary: List tools
      description: List all available tools
      operationId: listTools
      responses:
        '200':
          description: List of tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tool'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Tools
      summary: Create tool
      description: Create a new tool
      operationId: createTool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateToolRequest'
      responses:
        '201':
          description: Tool created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tools/{name}:
    get:
      tags:
        - Tools
      summary: Get tool
      description: Get tool by name
      operationId: getTool
      parameters:
        - name: name
          in: path
          required: true
          description: Tool name
          schema:
            type: string
      responses:
        '200':
          description: Tool details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Tools
      summary: Update tool
      description: Update an existing tool
      operationId: updateTool
      parameters:
        - name: name
          in: path
          required: true
          description: Tool name
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateToolRequest'
      responses:
        '200':
          description: Tool updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Tools
      summary: Delete tool
      description: Delete a tool by name
      operationId: deleteTool
      parameters:
        - name: name
          in: path
          required: true
          description: Tool name
          schema:
            type: string
      responses:
        '204':
          description: Tool deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /workflows/{workflow_id}/schedule:
    post:
      tags:
        - Workflows
      summary: Create or update workflow schedule
      operationId: createWorkflowSchedule
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cron_expression:
                  type: string
                timezone:
                  type: string
                enabled:
                  type: boolean
                next_run_at:
                  type: string
                  format: date-time
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowSchedule'

    get:
      tags:
        - Workflows
      summary: Get workflow schedule
      operationId: getWorkflowSchedule
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowSchedule'

    put:
      tags:
        - Workflows
      summary: Update workflow schedule
      operationId: updateWorkflowSchedule
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowSchedule'

    delete:
      tags:
        - Workflows
      summary: Delete workflow schedule
      operationId: deleteWorkflowSchedule
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Schedule deleted
...
  /admin/config:
    get:
      tags:
        - Admin
      summary: Get redacted config (admin only)
      operationId: getAdminConfig
      responses:
        '200':
          description: Redacted configuration
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '403':
          description: Forbidden
  /admin/diagnostics:
    get:
      tags:
        - Admin
      summary: Get diagnostics (DB, modules, version) (admin only)
      operationId: getAdminDiagnostics
      responses:
        '200':
          description: Diagnostics payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
                  database:
                    type: object
                  modules:
                    type: object
        '403':
          description: Forbidden
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API key authentication using Bearer token
  parameters:
    AgentId:
      name: id
      in: path
      required: true
      description: Agent UUID
      schema:
        type: string
        format: uuid
...
  schemas:
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Agent UUID
        name:
          type: string
          description: Agent name
        description:
          type: string
          nullable: true
          description: Agent description
        system_prompt:
          type: string
          description: System prompt for the agent
        model_name:
          type: string
          description: LLM model name
        memory_table:
          type: string
          nullable: true
          description: Memory table name
        enabled_tools:
          type: array
          items:
            type: string
          description: List of enabled tool names
        config:
          type: object
          additionalProperties: true
          description: Agent configuration
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - system_prompt
        - model_name
        - enabled_tools
        - config
        - created_at
        - updated_at
    CreateAgentRequest:
      type: object
      properties:
        name:
          type: string
          description: Agent name
        description:
          type: string
          nullable: true
          description: Agent description
        system_prompt:
          type: string
          description: System prompt for the agent
        model_name:
          type: string
          description: LLM model name
        memory_table:
          type: string
          nullable: true
          description: Memory table name
        enabled_tools:
          type: array
          items:
            type: string
          description: List of enabled tool names
        config:
          type: object
          additionalProperties: true
          description: Agent configuration (temperature, max_tokens, etc.)
      required:
        - name
        - system_prompt
        - model_name
        - enabled_tools
        - config
...
  responses:
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal Server Error"
            message: "An internal error occurred"
            code: 500
security:
  - BearerAuth: []